Improved connectivity to Tello.
To support multiple Tellos in a network, the
design of the commander had to be changed.
Now, Tellos can be added manually with ip-addresses.
To execute a command on a specific Tello, the ip-address
has to be given, otherwise, the first Tello will be taken or
any Tello with the default ip-address if no mapping exists.

The logger-design had to be changed to log errors while initializing
the Tello-connectivity.